feat: isolated the data layer creation logic and updated packages
