+++ example/express/package.json
@@ -19,7 +19,7 @@
     "cors": "^2.8.5",
     "express": "^4.17.3",
     "http-errors": "^2.0.0",
-    "mongoose": "^6.2.10",
+    "mongoose": "^6.3.0",
     "morgan": "^1.10.0",
     "swagger-ui-express": "^4.3.0"
   },
@@ -28,10 +28,10 @@
     "@types/express": "^4.17.13",
     "@types/http-errors": "^1.8.2",
     "@types/morgan": "^1.9.3",
-    "@types/node": "^17.0.23",
+    "@types/node": "^17.0.24",
     "@types/swagger-ui-express": "^4.1.3",
-    "@typescript-eslint/eslint-plugin": "^5.18.0",
-    "@typescript-eslint/parser": "^5.18.0",
+    "@typescript-eslint/eslint-plugin": "^5.19.0",
+    "@typescript-eslint/parser": "^5.19.0",
     "dotenv": "^16.0.0",
     "eslint": "^8.13.0",
     "eslint-config-prettier": "^8.5.0",
