+++ example/fastify/package.json
@@ -18,16 +18,16 @@
     "ajv": "^6",
     "fastify": "^3.28.0",
     "fastify-cors": "^6.0.3",
-    "fastify-swagger": "^5.1.0",
+    "fastify-swagger": "^5.1.1",
     "http-errors": "^2.0.0",
-    "mongoose": "^6.2.10",
+    "mongoose": "^6.3.0",
     "pino-pretty": "^7.6.1"
   },
   "devDependencies": {
     "@types/http-errors": "^1.8.2",
-    "@types/node": "^17.0.23",
-    "@typescript-eslint/eslint-plugin": "^5.18.0",
-    "@typescript-eslint/parser": "^5.18.0",
+    "@types/node": "^17.0.24",
+    "@typescript-eslint/eslint-plugin": "^5.19.0",
+    "@typescript-eslint/parser": "^5.19.0",
     "dotenv": "^16.0.0",
     "eslint": "^8.13.0",
     "eslint-config-prettier": "^8.5.0",
