+++ lib/src/functions/api/database.js
@@ -1,0 +1,159 @@
+const { platform } = require('os')
+const { promisify } = require('util')
+const exec = promisify(require('child_process').exec)
+const writeFile = require('../../utils/writeFile')
+
+module.exports = async ({ mongo = true, projectName }) => {
+  const createFoldersCommand = `mkdir ${projectName}/src/database \
+  ${projectName}/src/database/mongo \
+  ${projectName}/src/database/mongo/models \
+  ${projectName}/src/database/mongo/queries`
+
+  if (platform() === 'win32')
+    await exec(createFoldersCommand.replaceAll('/', '\\'))
+  else await exec(createFoldersCommand)
+
+  if (mongo) {
+    const database = {
+      index: {
+        content: "export * from './mongo'\n",
+        file: `${projectName}/src/database/index.ts`
+      },
+      mongo: {
+        index: {
+          content: `export * from './models'
+export * from './queries'
+`,
+          file: `${projectName}/src/database/mongo/index.ts`
+        }
+      },
+      'mongo/models': {
+        index: {
+          content: "export * from './user'\n",
+          file: `${projectName}/src/database/mongo/models/index.ts`
+        },
+        user: {
+          content: `import { model, Schema } from 'mongoose'
+
+const UserSchema = new Schema<UserDBO>(
+  {
+    lastName: {
+      required: true,
+      type: String
+    },
+    name: {
+      required: true,
+      type: String
+    }
+  },
+  {
+    timestamps: true,
+    versionKey: false,
+    toObject: {
+      transform: (_, ret) => {
+        ret.id = ret._id.toString()
+        delete ret._id
+      }
+    }
+  }
+)
+
+const UserModel = model<UserDBO>('users', UserSchema)
+
+export { UserModel }
+`,
+          file: `${projectName}/src/database/mongo/models/user.ts`
+        }
+      },
+      'mongo/queries': {
+        index: {
+          content: "export * from './user'\n",
+          file: `${projectName}/src/database/mongo/queries/index.ts`
+        },
+        user: {
+          content: `import { Document, Types } from 'mongoose'
+
+import { UserModel } from '..'
+import { UserDTO } from 'schemas'
+
+const userDBOtoDTO = (
+  userDBO: Document<unknown, unknown, UserDBO> &
+    UserDBO & {
+      _id: Types.ObjectId
+    }
+): UserDTO => ({
+  ...userDBO.toObject(),
+  createdAt: userDBO.createdAt.toISOString(),
+  updatedAt: userDBO.updatedAt.toISOString()
+})
+
+const store = async (userData: UserDTO): Promise<UserDTO> => {
+  const user = new UserModel(userData)
+
+  await user.save()
+
+  return userDBOtoDTO(user)
+}
+
+const remove = async (
+  id: string | null = null
+): Promise<UserDTO | number | null> => {
+  if (id) {
+    const removedUser = await UserModel.findByIdAndRemove(id)
+
+    if (!removedUser) return null
+
+    return userDBOtoDTO(removedUser)
+  }
+
+  return (await UserModel.deleteMany({})).deletedCount
+}
+
+const get = async (
+  id: string | null = null
+): Promise<UserDTO[] | UserDTO | null> => {
+  if (id) {
+    const user = await UserModel.findById(id)
+
+    return user ? userDBOtoDTO(user) : null
+  }
+
+  const users = await UserModel.find({})
+
+  return users.map(u => userDBOtoDTO(u))
+}
+
+const update = async (userData: UserDTO): Promise<UserDTO | null> => {
+  const { id, ...rest } = userData
+  const user = await UserModel.findByIdAndUpdate(id, rest, { new: true })
+
+  return user ? userDBOtoDTO(user) : null
+}
+
+export { store, remove, get, update }
+`,
+          file: `${projectName}/src/database/mongo/queries/user.ts`
+        }
+      }
+    }
+
+    await writeFile(database.index.file, database.index.content)
+    await writeFile(database.mongo.index.file, database.mongo.index.content)
+    await writeFile(
+      database['mongo/models'].index.file,
+      database['mongo/models'].index.content
+    )
+    await writeFile(
+      database['mongo/models'].user.file,
+      database['mongo/models'].user.content
+    )
+    await writeFile(
+      database['mongo/queries'].index.file,
+      database['mongo/queries'].index.content
+    )
+    await writeFile(
+      database['mongo/queries'].user.file,
+      database['mongo/queries'].user.content
+    )
+  }
+}
