+++ lib/src/functions/api/index.js
@@ -1,7 +1,8 @@
 const os = require('os')
 const util = require('util')
 const exec = util.promisify(require('child_process').exec)
-const writeFile = require('../utils/writeFile')
+const writeFile = require('../../utils/writeFile')
+const database = require('./database')
 
 /*
  * Express api:
@@ -103,12 +104,14 @@ const writeFile = require('../utils/writeFile')
  * @param {String} args.projectVersion
  * @param {String} args.email
  * @param {Boolean|undefined} args.fastify
+ * @param {Boolean|undefined} args.mongo
  */
 module.exports = async ({
   projectName,
   projectVersion,
   email,
-  fastify = false
+  fastify = false,
+  mongo = true
 }) => {
   const data = {
     '@types': {
@@ -132,128 +135,6 @@ export {}
 }
 `,
         file: `${projectName}/src/@types/models/user.d.ts`
-      }
-    },
-    database: {
-      index: {
-        content: "export * from './mongo'\n",
-        file: `${projectName}/src/database/index.ts`
-      }
-    },
-    'database/mongo': {
-      index: {
-        content: `export * from './models'
-export * from './queries'
-`,
-        file: `${projectName}/src/database/mongo/index.ts`
-      }
-    },
-    'database/mongo/models': {
-      index: {
-        content: "export * from './user'\n",
-        file: `${projectName}/src/database/mongo/models/index.ts`
-      },
-      user: {
-        content: `import { model, Schema } from 'mongoose'
-
-const UserSchema = new Schema<UserDBO>(
-  {
-    lastName: {
-      required: true,
-      type: String
-    },
-    name: {
-      required: true,
-      type: String
-    }
-  },
-  {
-    timestamps: true,
-    versionKey: false,
-    toObject: {
-      transform: (_, ret) => {
-        ret.id = ret._id.toString()
-        delete ret._id
-      }
-    }
-  }
-)
-
-const UserModel = model<UserDBO>('users', UserSchema)
-
-export { UserModel }
-`,
-        file: `${projectName}/src/database/mongo/models/user.ts`
-      }
-    },
-    'database/mongo/queries': {
-      index: {
-        content: "export * from './user'\n",
-        file: `${projectName}/src/database/mongo/queries/index.ts`
-      },
-      user: {
-        content: `import { Document, Types } from 'mongoose'
-
-import { UserModel } from '..'
-import { UserDTO } from 'schemas'
-
-const userDBOtoDTO = (
-  userDBO: Document<unknown, unknown, UserDBO> &
-    UserDBO & {
-      _id: Types.ObjectId
-    }
-): UserDTO => ({
-  ...userDBO.toObject(),
-  createdAt: userDBO.createdAt.toISOString(),
-  updatedAt: userDBO.updatedAt.toISOString()
-})
-
-const store = async (userData: UserDTO): Promise<UserDTO> => {
-  const user = new UserModel(userData)
-
-  await user.save()
-
-  return userDBOtoDTO(user)
-}
-
-const remove = async (
-  id: string | null = null
-): Promise<UserDTO | number | null> => {
-  if (id) {
-    const removedUser = await UserModel.findByIdAndRemove(id)
-
-    if (!removedUser) return null
-
-    return userDBOtoDTO(removedUser)
-  }
-
-  return (await UserModel.deleteMany({})).deletedCount
-}
-
-const get = async (
-  id: string | null = null
-): Promise<UserDTO[] | UserDTO | null> => {
-  if (id) {
-    const user = await UserModel.findById(id)
-
-    return user ? userDBOtoDTO(user) : null
-  }
-
-  const users = await UserModel.find({})
-
-  return users.map(u => userDBOtoDTO(u))
-}
-
-const update = async (userData: UserDTO): Promise<UserDTO | null> => {
-  const { id, ...rest } = userData
-  const user = await UserModel.findByIdAndUpdate(id, rest, { new: true })
-
-  return user ? userDBOtoDTO(user) : null
-}
-
-export { store, remove, get, update }
-`,
-        file: `${projectName}/src/database/mongo/queries/user.ts`
       }
     },
     network: {
@@ -558,7 +439,6 @@ Server.start()
       file: `${projectName}/src/index.ts`
     }
   }
-
   const expressData = {
     '@types/custom': {
       request: {
@@ -1378,7 +1258,6 @@ export { validatorCompiler }
       }
     }
   }
-
   const fastifyData = {
     network: {
       response: {
@@ -1857,10 +1736,6 @@ ${projectName}/src/network/routes/utils`
   const createFoldersCommands = `mkdir ${projectName}/src \
 ${projectName}/src/@types \
 ${projectName}/src/@types/models \
-${projectName}/src/database \
-${projectName}/src/database/mongo \
-${projectName}/src/database/mongo/models \
-${projectName}/src/database/mongo/queries \
 ${projectName}/src/network \
 ${projectName}/src/network/routes \
 ${projectName}/src/schemas \
@@ -1874,6 +1749,12 @@ ${fastify ? `${fastifyFolders}` : `${expressFolders}`}
     await exec(createFoldersCommands.replaceAll('/', '\\'))
   else await exec(createFoldersCommands)
 
+  // /database
+  database({
+    mongo,
+    projectName
+  })
+
   // /@types
   await writeFile(data['@types'].index.file, data['@types'].index.content)
 
@@ -1881,29 +1762,6 @@ ${fastify ? `${fastifyFolders}` : `${expressFolders}`}
   await writeFile(
     data['@types/models'].user.file,
     data['@types/models'].user.content
-  )
-
-  // /database
-  await writeFile(data.database.index.file, data.database.index.content)
-  await writeFile(
-    data['database/mongo'].index.file,
-    data['database/mongo'].index.content
-  )
-  await writeFile(
-    data['database/mongo/models'].index.file,
-    data['database/mongo/models'].index.content
-  )
-  await writeFile(
-    data['database/mongo/models'].user.file,
-    data['database/mongo/models'].user.content
-  )
-  await writeFile(
-    data['database/mongo/queries'].index.file,
-    data['database/mongo/queries'].index.content
-  )
-  await writeFile(
-    data['database/mongo/queries'].user.file,
-    data['database/mongo/queries'].user.content
   )
 
   // /schemas
